Make global bot instance optional so that we can initialise it later with a bot token.

+++ src/UI/Panels/MusicPanel.cpp
@@ -34,19 +34,19 @@ namespace Strawberry::Accoutrement
 
 	// clang-format off
 	wxBEGIN_EVENT_TABLE(MusicPanel, wxPanel)
-	EVT_BUTTON(Component::AddSongToDatabaseButton, MusicPanel::OnAddSong)
-	EVT_BUTTON(Component::EnqueueSongToPlaylistButton, MusicPanel::OnEnqueueSong)
-	EVT_BUTTON(Component::RemoveSongFromDatabaseButton, MusicPanel::OnRemoveFromDatabase)
-	EVT_BUTTON(Component::RemoveSongFromPlaylistButton, MusicPanel::OnRemoveSong)
-	EVT_BUTTON(Component::RenameSongButton, MusicPanel::OnRenameSong)
-	EVT_UPDATE_UI(wxID_ANY, MusicPanel::OnUpdate)
+					EVT_BUTTON(Component::AddSongToDatabaseButton, MusicPanel::OnAddSong)
+					EVT_BUTTON(Component::EnqueueSongToPlaylistButton, MusicPanel::OnEnqueueSong)
+					EVT_BUTTON(Component::RemoveSongFromDatabaseButton, MusicPanel::OnRemoveFromDatabase)
+					EVT_BUTTON(Component::RemoveSongFromPlaylistButton, MusicPanel::OnRemoveSong)
+					EVT_BUTTON(Component::RenameSongButton, MusicPanel::OnRenameSong)
+					EVT_UPDATE_UI(wxID_ANY, MusicPanel::OnUpdate)
 	wxEND_EVENT_TABLE();
 	// clang-format on
 
 
 	MusicPanel::MusicPanel(wxWindow* parent)
-		: wxPanel(parent)
-		, mEventReceiver(Bot::Get().GetPlaylist().Lock()->CreateEventReceiver())
+			: wxPanel(parent),
+			  mEventReceiver(Bot::Get() ? Bot::Get()->GetPlaylist().Lock()->CreateEventReceiver() : nullptr)
 	{
 		SetWindowStyle(wxSUNKEN_BORDER);
 
@@ -59,7 +59,8 @@ namespace Strawberry::Accoutrement
 		mSongDatabaseList->SetWindowStyle(wxLC_LIST);
 		for (int i = 0; i < SongDatabase::Get().GetNumSongs(); i++)
 		{
-			auto index = mSongDatabaseList->InsertItem(mSongDatabaseList->GetItemCount(), SongDatabase::Get().GetSong(i).GetTitle());
+			auto index = mSongDatabaseList->InsertItem(mSongDatabaseList->GetItemCount(),
+													   SongDatabase::Get().GetSong(i).GetTitle());
 			mSongDatabaseList->SetItemPtrData(index, i);
 		}
 		sizer->Add(mSongDatabaseList, {1, 0}, {1, 1}, wxEXPAND | wxALL, 5);
@@ -78,7 +79,8 @@ namespace Strawberry::Accoutrement
 		sizer->Add(songListButtons, {3, 0}, {1, 1}, wxALIGN_CENTER_HORIZONTAL, 5);
 
 		auto playlistButtons = new wxGridBagSizer(5, 5);
-		playlistButtons->Add(new wxButton(this, Component::RemoveSongFromPlaylistButton, "Remove"), {0, 0}, {1, 1}, wxEXPAND | wxALL, 5);
+		playlistButtons->Add(new wxButton(this, Component::RemoveSongFromPlaylistButton, "Remove"), {0, 0}, {1, 1},
+							 wxEXPAND | wxALL, 5);
 		playlistButtons->Add(new wxButton(this, wxID_ANY, "Move Up"), {0, 1}, {1, 1}, wxEXPAND | wxALL, 5);
 		playlistButtons->Add(new wxButton(this, wxID_ANY, "Move Down"), {1, 1}, {1, 1}, wxEXPAND | wxALL, 5);
 		playlistButtons->Add(new wxButton(this, wxID_ANY, "Shuffle"), {0, 2}, {1, 1}, wxEXPAND | wxALL, 5);
@@ -95,24 +97,28 @@ namespace Strawberry::Accoutrement
 
 	void MusicPanel::OnUpdate(wxUpdateUIEvent& event)
 	{
-		auto playlistMessage = mEventReceiver->Read();
-		if (!playlistMessage) return;
+		if (mEventReceiver)
+		{
+			auto playlistMessage = mEventReceiver->Read();
+			if (!playlistMessage) return;
 
-		if (auto songAdded = playlistMessage->Value<Codec::Audio::Playlist::SongAddedEvent>())
-		{
-			Song song = std::any_cast<Song>(songAdded->associatedData);
-			mPlaylistView->InsertItem(songAdded->index, song.GetTitle());
-		}
-		else if (auto songRemoved = playlistMessage->Value<Codec::Audio::Playlist::SongRemovedEvent>())
-		{
-			Core::Assert(songRemoved->index < mPlaylistView->GetItemCount());
-			mPlaylistView->DeleteItem(songRemoved->index);
-		}
-		else if (auto songBegan = playlistMessage->Value<Codec::Audio::Playlist::SongBeganEvent>())
-		{
-			for (int i = 0; i < mPlaylistView->GetItemCount(); i++)
+			if (auto songAdded = playlistMessage->Value<Codec::Audio::Playlist::SongAddedEvent>())
 			{
-				mPlaylistView->SetItemBackgroundColour(i, i == songBegan->index ? wxColor(32, 128, 32, 255) : mPlaylistView->GetBackgroundColour());
+				Song song = std::any_cast<Song>(songAdded->associatedData);
+				mPlaylistView->InsertItem(songAdded->index, song.GetTitle());
+			}
+			else if (auto songRemoved = playlistMessage->Value<Codec::Audio::Playlist::SongRemovedEvent>())
+			{
+				Core::Assert(songRemoved->index < mPlaylistView->GetItemCount());
+				mPlaylistView->DeleteItem(songRemoved->index);
+			}
+			else if (auto songBegan = playlistMessage->Value<Codec::Audio::Playlist::SongBeganEvent>())
+			{
+				for (int i = 0; i < mPlaylistView->GetItemCount(); i++)
+				{
+					mPlaylistView->SetItemBackgroundColour(i, i == songBegan->index ? wxColor(32, 128, 32, 255)
+																					: mPlaylistView->GetBackgroundColour());
+				}
 			}
 		}
 	}
@@ -123,20 +129,21 @@ namespace Strawberry::Accoutrement
 		wxFileDialog fileDialog(this, "Choose a song file...");
 		fileDialog.SetWindowStyle(wxFD_OPEN | wxFD_MULTIPLE | wxFD_FILE_MUST_EXIST);
 		auto dialogResult = fileDialog.ShowModal();
-		if (dialogResult == wxID_CANCEL) { return; }
+		if (dialogResult == wxID_CANCEL)
+		{ return; }
 		else if (dialogResult == wxID_OK)
 		{
 			wxArrayString paths;
 			fileDialog.GetPaths(paths);
-			for (const auto& path : paths)
+			for (const auto& path: paths)
 			{
 				auto fullPath = std::filesystem::absolute(std::string(path));
-				auto song     = Song::FromFile(fullPath);
+				auto song = Song::FromFile(fullPath);
 				if (!song) return;
 
-				auto        songIndex = SongDatabase::Get().AddSong(song.Value());
-				std::string title     = song.Value().GetTitle();
-				auto        index     = mSongDatabaseList->InsertItem(mSongDatabaseList->GetItemCount(), title);
+				auto songIndex = SongDatabase::Get().AddSong(song.Value());
+				std::string title = song.Value().GetTitle();
+				auto index = mSongDatabaseList->InsertItem(mSongDatabaseList->GetItemCount(), title);
 				mSongDatabaseList->SetItemPtrData(index, songIndex);
 			}
 		}
@@ -145,57 +152,67 @@ namespace Strawberry::Accoutrement
 
 	void MusicPanel::OnEnqueueSong(wxCommandEvent& event)
 	{
-		size_t selectedSongIndex = mSongDatabaseList->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
-		if (selectedSongIndex == -1) return;
+		if (Bot::Get())
+		{
+			size_t selectedSongIndex = mSongDatabaseList->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
+			if (selectedSongIndex == -1) return;
 
-		wxListItem selectedSongItem;
-		selectedSongItem.SetId(selectedSongIndex);
-		if (!mSongDatabaseList->GetItem(selectedSongItem)) return;
+			wxListItem selectedSongItem;
+			selectedSongItem.SetId(selectedSongIndex);
+			if (!mSongDatabaseList->GetItem(selectedSongItem)) return;
 
-		selectedSongIndex = selectedSongItem.GetData();
+			selectedSongIndex = selectedSongItem.GetData();
 
-		auto playlist     = Bot::Get().GetPlaylist().Lock();
-		auto song         = SongDatabase::Get().GetSong(selectedSongIndex);
-		playlist->EnqueueFile(song.GetPath(), song).Unwrap();
+			auto playlist = Bot::Get()->GetPlaylist().Lock();
+			auto song = SongDatabase::Get().GetSong(selectedSongIndex);
+			playlist->EnqueueFile(song.GetPath(), song).Unwrap();
+		}
 	}
 
 
 	void MusicPanel::OnRemoveSong(wxCommandEvent& event)
 	{
-		auto index = mPlaylistView->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
-		if (index != -1) { Bot::Get().GetPlaylist().Lock()->RemoveTrack(index); }
+		if (Bot::Get())
+		{
+			auto index = mPlaylistView->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
+			if (index != -1)
+				Bot::Get()->GetPlaylist().Lock()->RemoveTrack(index);
+		}
 	}
 
 
 	void MusicPanel::OnRenameSong(wxCommandEvent& event)
 	{
-		long item = mSongDatabaseList->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
-		if (item != -1)
+		if (Bot::Get())
 		{
-			auto  songID    = mSongDatabaseList->GetItemData(item);
-			auto& song      = SongDatabase::Get().GetSong(songID);
-			auto  strDialog = new wxTextEntryDialog(this, "Choose a new name:");
-			strDialog->SetValue(song.GetTitle());
-			if (strDialog->ShowModal() == wxID_OK)
+			long item = mSongDatabaseList->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
+			if (item != -1)
 			{
-				std::string title(strDialog->GetValue());
-
-				song.SetTitle(title);
-				mSongDatabaseList->SetItem(item, 0, title);
-
-				auto playlist = Bot::Get().GetPlaylist().Lock();
-				for (int i = 0; i < playlist->Length(); i++)
+				auto songID = mSongDatabaseList->GetItemData(item);
+				auto& song = SongDatabase::Get().GetSong(songID);
+				auto strDialog = new wxTextEntryDialog(this, "Choose a new name:");
+				strDialog->SetValue(song.GetTitle());
+				if (strDialog->ShowModal() == wxID_OK)
 				{
-					Song playlistSong = playlist->GetTrackAssociatedData<Song>(i);
-					if (song.GetPath() == playlistSong.GetPath())
-					{
-						playlistSong.SetTitle(title);
-						playlist->SetTrackAssociatedData(i, playlistSong);
-						mPlaylistView->SetItem(i, 0, title);
-					}
-				}
+					std::string title(strDialog->GetValue());
 
-				Layout();
+					song.SetTitle(title);
+					mSongDatabaseList->SetItem(item, 0, title);
+
+					auto playlist = Bot::Get()->GetPlaylist().Lock();
+					for (int i = 0; i < playlist->Length(); i++)
+					{
+						Song playlistSong = playlist->GetTrackAssociatedData<Song>(i);
+						if (song.GetPath() == playlistSong.GetPath())
+						{
+							playlistSong.SetTitle(title);
+							playlist->SetTrackAssociatedData(i, playlistSong);
+							mPlaylistView->SetItem(i, 0, title);
+						}
+					}
+
+					Layout();
+				}
 			}
 		}
 	}
